Split cells feature + code migration (#8)
* Basic split setup

* split seems to be working

* Add proper perturb and split probability

* Minor config changes

* Move 3d to separate folder and add first 20 frames
