+++ 3d/examples/3d/initial.csv
@@ -1,0 +1,7 @@
+file,name,x,y,z,radius,z_scaling,split_alpha,opacity
+frame000.tif,e3d03428900d4123ae24ac57ad695c8d,111.9472222222222,202.79444444444442,12,14.5,6.952,None,None
+frame000.tif,12345679342524354354502034985234,130.94722222222222,160.79444444444442,18,30,6.952,None,None
+frame000.tif,e9077677575842b1a2925729fbcfb3a5,303.60555555555555,155.31944444444443,18,28.040149785619903,6.952,None,None
+frame000.tif,1f2ed10d323c4cb288424e988893788f,395.625,294.22777777777776,21,29.60152023123137,6.952,None,None
+frame000.tif,1f89abf484c94c498a23cad71ebee0cb,199.86388888888888,334.6694444444444,19,31.12073263919087,6.952,None,None
+frame000.tif,8cbdf86d308d4599936e7fdbc23375f5,338.7722222222222,255.54444444444442,32,18.439088914585774,6.952,None,None
