--- examples/3d/config.yaml
@@ -1,70 +1,0 @@
-# Global settings
-cellType: sphere
-#pixelsPerMicron: 1
-#framesPerSecond: 1
-
-# Cell settings
-cell:
-  x:
-    prob: 0.35
-    mu: 0
-    sigma: 10
-  y:
-    prob: 0.35
-    mu: 0
-    sigma: 10
-  z:
-    prob: 0.25
-    mu: 0
-    sigma: 10
-  radius:
-    prob: 0.1
-    mu: 0
-    sigma: 3
-  minRadius: 4
-  maxRadius: 24
-
-# Simulation settings
-simulation:
-  iterations_per_cell: 1
-  background_color: 0.376
-  cell_color: 0.496
-  padding: 0
-  z_scaling: 6.925
-  blur_sigma: 2
-
-# Probability of changes
-#prob:
-#  perturbation: 0.7
-#  combine: 0.1
-#  split: 0.12
-#  camera_shift: 0.0
-#  opacity_diffraction_offset: 0.05
-#  background_offset: 0.05
-
-# Camera shift settings
-#camera:
-#  modification_x_sigma: 0
-#  modification_y_sigma: 0
-
-#global_optimizer_window_size: 1
-#pbad_max_size: 1000
-#auto_temp_scheduler_iteration_per_cell: 500
-#output_format: png
-#output_quality: 80
-#residual_vmin: -0.3
-#residual_vmax: 0.3
-#split_cost: 0
-#combine_cost: 0
-#overlap_cost: 1
-#cell_importance: 1
-#background_offset_mu: 0
-#background_offset_sigma: 0.01
-#cell_brightness_mu: 0
-#cell_brightness_sigma: 0.01
-#opacity_offset_mu: 0
-#opacity_offset_sigma: 0.01
-#diffraction_strength_offset_mu: 0
-#diffraction_strength_offset_sigma: 0.01
-#diffraction_sigma_offset_mu: 0
-#diffraction_sigma_offset_sigma: 0.01
