--- examples/3d/global_optimizer_config.json
@@ -1,95 +1,0 @@
-{
-    // Global settings
-    "global.cellType": "sphere",
-    "global.pixelsPerMicron": 1.0,
-    "global.framesPerSecond": 1.0,
-
-    // Bacilli settings
-    "sphere.maxSpeed":  200.0,  // microns per second
-    "sphere.maxSpin":    1.57, // radians per second
-    "sphere.minGrowth": -2.0,  // microns
-    "sphere.maxGrowth": 43.0,  // microns
-    "sphere.minWidth":   4.0,  // microns
-    "sphere.maxWidth":   800.0,  // microns
-    "sphere.minLength": 14.0,  // microns
-    "sphere.maxLength": 800.0,  // microns
-    "sphere.distanceCostDivisor": 24,  // microns
-    
-    // simulation settings. NOTE: Image are converted to grayscale float32. 
-    // All settings should be floats.
-    "simulation":
-    {
-        "image.type": "To be Set", 
-        "background.color": 0, // 0.39 for gary scale image
-        "cell.color" : 1, // 0.15 for gray scale image
-        "light.diffraction.sigma": "auto",  //11
-        "light.diffraction.strength": "auto", //0.6
-        "light.diffraction.truncate": 1,
-        "cell.opacity": "auto",
-        "padding": 0,
-        "image.slices": 33
-    },
-    
-    "prob.perturbation": 0.70,
-    "prob.combine": 0.1,
-    "prob.split" : 0.1,
-    "prob.camera_shift": 0.0,
-    "prob.opacity_diffraction_offset": 0.05,
-    "prob.background_offset": 0.05, //This should be set to 0 if the algorithm runs in binary version
-    "perturbation":
-    {
-        "prob.x": 0.35,
-        "prob.y": 0.35,
-        "prob.z": 0.25,
-        "prob.width":0.1,
-        "prob.length":0.2,
-        "prob.rotation":0.2,
-        "modification.x.mu":0.0,
-        "modification.y.mu":0.0,
-        "modification.width.mu":0.0,
-        "modification.length.mu":0.0,
-        "modification.rotation.mu":0.0,
-        "modification.x.sigma":10.0,
-        "modification.y.sigma":10.0,
-        "modification.width.sigma":0.1,
-        "modification.length.sigma":1.0,
-        "modification.rotation.sigma":0.2,
-        "prob.opacity": 0.2,
-        "modification.opacity.mu":0.0,
-        "modification.opacity.sigma":0.001
-    },
-
-    "camera":
-    {
-        "modification.x.sigma": 0,
-        "modification.y.sigma": 0
-    },
-
-    "global_optimizer.window_size": 1,
-    "pbad_max_size": 1000, // must be greater than 0
-    "iteration_per_cell": 0,
-    "auto_temp_scheduler.iteration_per_cell": 500,
-    "output.format": "png", //must be png or jpeg. If jpeg is chosen, then the output image quality can be set.
-    "output.quality": 80, // range [0, 100]
-    //residual.vmin must equal -residual.vmax
-    "residual.vmin": -0.3,
-    "residual.vmax": 0.3,
-    //objective parameter
-    "split.cost": 0,
-
-    "combine.cost": 0,
-    "overlap.cost": 1,
-    //cell.importance = importance of synthetic cell region/importance of synthetic diffraction region
-    "cell.importance": 1,
-
-    "background_offset.mu": 0.0,
-    "background_offset.sigma": 0.01,
-    "cell_brightness.mu": 0.0,
-    "cell_brightness.sigma": 0.01,
-    "opacity_offset.mu": 0,
-    "opacity_offset.sigma": 0.01,
-    "diffraction_strength_offset.mu": 0,
-    "diffraction_strength_offset.sigma": 0.01,
-    "diffraction_sigma_offset.mu":  0,
-    "diffraction_sigma_offset.sigma": 0.01
-}
