--- examples/3d/initial.csv
@@ -1,9 +1,0 @@
-file,name,x,y,width,length,rotation,split_alpha,opacity,z
-image0000.jpg,b000,300,200,60.373756293,20.07513143,1.611157981,None,None,15
-image0000.jpg,b001,300,200,60.373756293,20.07513143,1.611157981,None,None,10
-image0000.jpg,b010,100,300,60.600692808,18.67014573,2.016786596,None,None,30
-image0000.jpg,b100,300,300,60.585344495,17.44037879,2.054069441,None,None,-3
-image0000.jpg,b101,300,300,60.728826589,18.04443386,1.535245464,None,None,-10
-image0000.jpg,b011,400,100,60.728826589,18.04443386,1.535245464,None,None,-13
-image0000.jpg,b110,400,100,60.728826589,18.04443386,1.535245464,None,None,-50
-image0000.jpg,b111,400,100,60.728826589,18.04443386,1.535245464,None,None,15
