+++ mypy.ini
@@ -1,5 +1,5 @@
 [mypy]
 plugins = numpy.typing.mypy_plugin
-mypy_path = src
+mypy_path = 3d
 check_untyped_defs = True
 ignore_missing_imports = True
