--- src/global_optimization/Cells/__init__.py
@@ -1,5 +1,0 @@
-from .Cell import Cell
-from .Bacilli import BacilliConfig, Bacilli
-from .Sphere import SphereConfig, Sphere
-from .CellFactory import CellFactory
-
