--- src/global_optimization/Config/__init__.py
@@ -1,2 +1,0 @@
-from .Config import load_config
-from .ConfigTypes import BaseConfig, SimulationConfig, CellConfig
