--- src/global_optimization/utils.py
@@ -1,3 +1,0 @@
-def gerp(a, b, t):
-    """Geometric interpolation"""
-    return a * (b / a) ** t
